3-claude-desktop/pyproject.toml
[project]
name = "claude-desktop-mcp-demo"
version = "0.1.0"
description = ""
authors = [
    {name = "Jody Hunt",email = "jody.hunt@cyberark.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastmcp (>=2.11.3,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
